Win32-network: clean package
Removed:
* events api
* synchronous file handle api
